requirements.txt
tensorflow-cpu
psutil
GPUtil
scikit-learn
flask
celery
requests
pyelftools
